refactor; move all user notifications related code
to de.metas.notifications

https://github.com/metasfresh/metasfresh/issues/4002
