+++ de.metas.adempiere.adempiere/base/src/main/java/de/metas/notification/IUserNotificationsConfigRepository.java
@@ -1,9 +1,6 @@
-package org.adempiere.user.api;
+package de.metas.notification;
 
-import org.adempiere.util.Check;
-
-import lombok.NonNull;
-import lombok.Value;
+import org.adempiere.util.ISingletonService;
 
 /*
  * #%L
@@ -15,31 +12,19 @@ import lombok.Value;
  * it under the terms of the GNU General Public License as
  * published by the Free Software Foundation, either version 2 of the
  * License, or (at your option) any later version.
- *
+ * 
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
- *
+ * 
  * You should have received a copy of the GNU General Public
  * License along with this program. If not, see
  * <http://www.gnu.org/licenses/gpl-2.0.html>.
  * #L%
  */
 
-@Value
-public class NotificationGroupName
+public interface IUserNotificationsConfigRepository extends ISingletonService
 {
-	public static NotificationGroupName of(final String valueAsString)
-	{
-		return new NotificationGroupName(valueAsString);
-	}
-
-	private final String valueAsString;
-
-	private NotificationGroupName(@NonNull final String valueAsString)
-	{
-		Check.assumeNotEmpty(valueAsString, "valueAsString shall not be empty");
-		this.valueAsString = valueAsString;
-	}
+	UserNotificationsConfig getByUserId(int adUserId);
 }
