+++ de.metas.adempiere.adempiere/base/src/main/java/de/metas/notification/NotificationGroupName.java
@@ -1,14 +1,8 @@
-package org.adempiere.user.api;
-
-import java.util.List;
+package de.metas.notification;
 
 import org.adempiere.util.Check;
 
-import com.google.common.collect.ImmutableList;
-
-import lombok.Builder;
 import lombok.NonNull;
-import lombok.Singular;
 import lombok.Value;
 
 /*
@@ -21,12 +15,12 @@ import lombok.Value;
  * it under the terms of the GNU General Public License as
  * published by the Free Software Foundation, either version 2 of the
  * License, or (at your option) any later version.
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public
  * License along with this program. If not, see
  * <http://www.gnu.org/licenses/gpl-2.0.html>.
@@ -34,18 +28,18 @@ import lombok.Value;
  */
 
 @Value
-public class RoleNotificationsConfig
+public class NotificationGroupName
 {
-	private final int roleId;
-	private List<UserNotificationsGroup> notificationGroups;
-
-	@Builder
-	public RoleNotificationsConfig(
-			final int roleId,
-			@NonNull @Singular final ImmutableList<UserNotificationsGroup> notificationGroups)
+	public static NotificationGroupName of(final String valueAsString)
 	{
-		Check.assumeGreaterOrEqualToZero(roleId, "roleId");
-		this.roleId = roleId;
-		this.notificationGroups = notificationGroups;
+		return new NotificationGroupName(valueAsString);
+	}
+
+	private final String valueAsString;
+
+	private NotificationGroupName(@NonNull final String valueAsString)
+	{
+		Check.assumeNotEmpty(valueAsString, "valueAsString shall not be empty");
+		this.valueAsString = valueAsString;
 	}
 }
