+++ de.metas.adempiere.adempiere/base/src/main/java/org/adempiere/user/api/IUserBL.java
@@ -112,9 +112,4 @@ public interface IUserBL extends ISingletonService
 			throw new AdempiereException("User cannot send emails: " + errmsg.translate(Env.getAD_Language(Env.getCtx())));
 		}
 	}
-
-	UserNotificationsConfig getUserNotificationsConfig(int adUserId);
-
-	RoleNotificationsConfig getRoleNotificationsConfig(int adRoleId);
-
 }
