+++ de.metas.adempiere.adempiere/base/src/main/java/org/adempiere/user/api/IUserDAO.java
@@ -79,8 +79,4 @@ public interface IUserDAO extends ISingletonService
 	 * @return AD_User_IDs
 	 */
 	List<Integer> retrieveSystemUserIds();
-
-	UserNotificationsConfig getUserNotificationsConfig(int adUserId);
-
-	RoleNotificationsConfig getRoleNotificationsConfig(int adRoleId);
 }
