+++ de.metas.adempiere.adempiere/base/src/main/java/org/adempiere/user/api/impl/UserBL.java
@@ -42,8 +42,6 @@ import org.adempiere.service.IClientDAO;
 import org.adempiere.service.ISysConfigBL;
 import org.adempiere.user.api.IUserBL;
 import org.adempiere.user.api.IUserDAO;
-import org.adempiere.user.api.RoleNotificationsConfig;
-import org.adempiere.user.api.UserNotificationsConfig;
 import org.adempiere.util.Check;
 import org.adempiere.util.Services;
 import org.compiere.model.I_AD_Org;
@@ -485,17 +483,4 @@ public class UserBL implements IUserBL
 		final I_AD_User user = Services.get(IUserDAO.class).retrieveUser(adUserId);
 		return checkCanSendEMail(user);
 	}
-
-	@Override
-	public UserNotificationsConfig getUserNotificationsConfig(final int adUserId)
-	{
-		return Services.get(IUserDAO.class).getUserNotificationsConfig(adUserId);
-	}
-
-	@Override
-	public RoleNotificationsConfig getRoleNotificationsConfig(final int adRoleId)
-	{
-		return Services.get(IUserDAO.class).getRoleNotificationsConfig(adRoleId);
-	}
-
 }
